move MCMCSampler baed classes to mcmc_sampler_base module + cleanup
