+++ qumcmc/quantum_mcmc_routines.py
@@ -227,39 +227,3 @@ def quantum_enhanced_mcmc(
     return mcmc_chain
 
 ####
-## decorator class for running quantum_enhanced_mcmc() 
-from dataclasses import dataclass
-from .mcmc_sampler_base import mcmc_sampler
-@dataclass
-class qumcmc_sampler(mcmc_sampler) : 
-
-    def __init__(self,
-        n_hops: int ,
-        model: IsingEnergyFunction,
-        mixer: Mixer,
-        gamma: GammaType,
-        initial_state: Optional[str] = None,
-        temperature: float = 1,
-        delta_time=0.8,
-        verbose: bool = False,
-        name: str = "Q-MCMC") -> None : 
-        
-        self.mixer = mixer 
-        self.gamma = gamma 
-        self.delta_time = delta_time
-        super().__init__(n_hops, model, initial_state, temperature, verbose, name)
-
-    # super.__init__(n_hops, model, initial_state, temperature, verbose, name)
-    def __repr__(self):
-        return f"Quantum MCMCSampler: {self.name} \n ------------------- Iterations : {self.n_hops} \n Model : {self.model} \n Mixer : {self.mixer} \n initial-state : {self.initial_state} \ngamma : {self.gamma} | temp : {self.temperature} | evolution-time : {self.delta_time}\n  " 
-        # pass 
-    def run(self): 
-        return quantum_enhanced_mcmc(n_hops= self.n_hops ,
-                                     model= self.model , 
-                                     mixer = self.mixer , 
-                                     gamma= self.gamma ,
-                                     initial_state= self.initial_state , 
-                                     temperature= self.temperature , 
-                                     delta_time= self.delta_time , 
-                                     verbose= self.verbose , 
-                                     name = self.name)
